+++ src/utils/element.ts
@@ -318,7 +318,7 @@ export const completeSize = (size: number | string, unit = 'px') => {
     return size.toString() + unit
   } else if (
     isValueType<string>(size, 'String') &&
-    APP_REGEX.validCSSUnit.test(size)
+    APP_REGEX.cssUnit.test(size)
   ) {
     return size
   } else {